[Feat] 토큰 발급 및 대기열 조회 로직 추가
