+++ src/main/java/kr/hhplus/be/server/application/TokenService.java
@@ -1,0 +1,59 @@
+package kr.hhplus.be.server.application;
+
+import java.util.UUID;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import kr.hhplus.be.server.domain.token.Token;
+import kr.hhplus.be.server.domain.token.TokenInfo;
+import kr.hhplus.be.server.domain.token.TokenRepository;
+import kr.hhplus.be.server.domain.token.TokenStatus;
+import kr.hhplus.be.server.domain.user.User;
+import kr.hhplus.be.server.domain.user.UserRepository;
+import kr.hhplus.be.server.global.error.CustomErrorCode;
+import kr.hhplus.be.server.global.error.CustomException;
+import kr.hhplus.be.server.interfaces.token.TokenSearchInfo;
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class TokenService {
+
+	private final UserRepository userRepository;
+	private final TokenRepository tokenRepository;
+
+	@Transactional
+	public TokenInfo issueToken(long userId) {
+
+		User user = userRepository.findById(userId)
+			.orElseThrow(() -> new CustomException(CustomErrorCode.NOT_FOUND_USER));
+
+		Token issueToken = tokenRepository.findByUserId(userId).orElseGet(() -> {
+			return Token.createToken(user, TokenStatus.WAITING, UUID.randomUUID().toString());
+		});
+
+		return TokenInfo.from(issueToken);
+
+	}
+
+	@Transactional(readOnly = true)
+	public TokenSearchInfo searchToken(long userId) {
+		userRepository.findById(userId)
+			.orElseThrow(() -> new CustomException(CustomErrorCode.NOT_FOUND_USER));
+
+		Token token = tokenRepository.findByUserId(userId)
+			.orElseThrow(() -> new CustomException(CustomErrorCode.NOT_FOUND_TOKEN));
+
+		// 현재 ACTIVE 토큰 수 조회
+		long activeTokenCount = tokenRepository.countByStatus(TokenStatus.ACTIVE);
+
+		// 대기 순위를 조회 후 도메인 메서드에 외부 정보 함께 전달
+		int waitingRank = tokenRepository.getWaitingRank(token.getId());
+		token.checkAndActivate(waitingRank, activeTokenCount);
+
+		return TokenSearchInfo.from(token);
+
+	}
+
+}
