+++ src/main/java/kr/hhplus/be/server/global/error/CustomErrorCode.java
@@ -29,7 +29,8 @@ public enum CustomErrorCode {
 	NOT_FOUND_RESERVATION(HttpStatus.NOT_FOUND, "404", "예약 정보를 찾을 수 없습니다"),
 	NOT_FOUND_CONCERT_SEAT(HttpStatus.NOT_FOUND, "404", "좌석 정보를 찾을 수 없습니다"),
 	INVALID_PAYMENT_AMOUNT(HttpStatus.BAD_REQUEST, "400", "유효하지 않은 결제 금액입니다"),
-	INVALID_STATUS(HttpStatus.INTERNAL_SERVER_ERROR, "500", "유효한 상태가 아닙니다");
+	INVALID_STATUS(HttpStatus.INTERNAL_SERVER_ERROR, "500", "유효한 상태가 아닙니다"),
+	NOT_FOUND_TOKEN(HttpStatus.NOT_FOUND, "404", "토큰을 찾을 수 없습니다");
 
 	private final HttpStatus httpStatus;
 	private final String code;
