+++ src/main/java/kr/hhplus/be/server/interfaces/token/TokenController.java
@@ -1,8 +1,5 @@
 package kr.hhplus.be.server.interfaces.token;
 
-import java.time.LocalDateTime;
-import java.util.UUID;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,8 +7,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import kr.hhplus.be.server.domain.token.TokenService;
-import kr.hhplus.be.server.domain.token.TokenStatus;
+import kr.hhplus.be.server.application.TokenService;
 import kr.hhplus.be.server.global.config.swagger.TokenApi;
 import lombok.RequiredArgsConstructor;
 
@@ -37,19 +33,9 @@ public class TokenController implements TokenApi {
 	 *  차례가 되어 토큰이 활성화 된 부분을 확인
 	 */
 	@GetMapping("/{userId}")
-	public ResponseEntity<TokenResponse> searchToken(@PathVariable(name = "userId") long userId) {
-
-		// 실제 서비스의 호출 없이 , 고정 응답을 반환합니다.
-		TokenSearchInfo mockInfo = TokenSearchInfo.builder()
-			.tokenValue(UUID.randomUUID().toString())
-			.status(TokenStatus.ACTIVE)
-			.userId(userId)
-			.createdAt(LocalDateTime.now())
-			.updateAt(LocalDateTime.now().plusMinutes(2))
-			.build();
-
-		TokenResponse response = new TokenResponse("대기열 순서 조회 성공", mockInfo);
-		return ResponseEntity.ok().body(response);
+	public ResponseEntity<TokenSearchResponse> searchToken(@PathVariable(name = "userId") long userId) {
+		return ResponseEntity.ok()
+			.body(TokenSearchResponse.from("대기열 조회 성공", tokenService.searchToken(userId)));
 	}
 
 }
